PMM-6624 Fixing mongodb 4 4 issue with library  (#350)
* Remove UI tests

* PMM-6624 Adding missing library in case for mongo 4.4

+++ pmm-tests/pmm-framework.sh
@@ -1112,7 +1112,11 @@ add_clients(){
     ADDCLIENTS_COUNT=$(echo "${i}" | sed 's|[^0-9]||g')
     if  [[ "${CLIENT_NAME}" == "mo" && -z $MONGOMAGIC ]]; then
       rm -rf $BASEDIR/data
-      sudo ln -s /usr/lib64/liblzma.so.5.2.2 /usr/lib64/liblzma.so.0
+      if [ -f /usr/lib64/liblzma.so.5.0.99 ]; then
+        sudo ln -s /usr/lib64/liblzma.so.5.0.99 /usr/lib64/liblzma.so.0
+      else
+        sudo ln -s /usr/lib64/liblzma.so.5.2.2 /usr/lib64/liblzma.so.0
+      fi
       for k in `seq 1  ${REPLCOUNT}`;do
         PSMDB_PORT=$(( (RANDOM%21 + 10) * 1001 ))
         PSMDB_PORTS+=($PSMDB_PORT)
@@ -1495,7 +1499,11 @@ add_clients(){
       echo ${BASEDIR}
       
       ##Missing Library for 4.4
-      sudo ln -s /usr/lib64/liblzma.so.5.2.2 /usr/lib64/liblzma.so.0
+      if [ -f /usr/lib64/liblzma.so.5.0.99 ]; then
+        sudo ln -s /usr/lib64/liblzma.so.5.0.99 /usr/lib64/liblzma.so.0
+      else
+        sudo ln -s /usr/lib64/liblzma.so.5.2.2 /usr/lib64/liblzma.so.0
+      fi
       
       if [[ "$with_sharding" == "1" ]]; then
         if [ "$mo_version" == "3.6" ]; then
